fix: add support for list type data in field stats
current: stats for list type fields show UNSUPPORTED
change: add support for the list type fields

specially useful in otel metrics where `data_point_bucket_counts` and `data_point_explicit_bounds`
are left as arrays and not considered for flattening
