+++ src/storage/mod.rs
@@ -36,6 +36,7 @@ use chrono::Utc;
 use std::fmt::Debug;
 
 mod azure_blob;
+pub mod field_stats;
 mod localfs;
 mod metrics_layer;
 pub mod object_storage;
